+++ cmd/rabtap/main.go
@@ -141,8 +141,10 @@ func startCmdPublish(ctx context.Context, args CommandLineArgs) {
 		fixedDelay: args.Delay,
 		speed:      args.Speed,
 		tlsConfig:  getTLSConfig(args.InsecureTLS, args.TLSCertFile, args.TLSKeyFile, args.TLSCaFile),
+		mandatory:  args.Mandatory,
+		confirms:   args.Confirms,
 		readerFunc: readerFunc})
-	failOnError(err, "error publishing message", os.Exit)
+	failOnError(err, "publish", os.Exit)
 }
 
 func startCmdSubscribe(ctx context.Context, args CommandLineArgs) {